requirements.txt
aiodns==3.2.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==7.1.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
av==12.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.4
ctransformers==0.2.27
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
effdet==0.4.1
einops==0.8.0
emoji==2.10.1
et-xmlfile==1.1.0
eva-decord==0.6.1
exceptiongroup==1.2.0
fastapi==0.109.2
fastembed==0.1.3
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
fvcore==0.1.5.post20221221
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.25
langchain-text-splitters==0.2.0
langdetect==1.0.9
langsmith==0.1.5
layoutparser==0.3.4
llama-cloud==0.0.6
llama-index==0.10.52
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.52.post1
llama-index-embeddings-fastembed==0.1.3
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.1
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.23
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-qdrant==0.2.10
llama-parse==0.4.4
llama_cpp_python==0.2.44
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.35.7
opencv-python==4.9.0.80
openpyxl==3.1.2
orjson==3.10.3
packaging==23.2
pandas==2.2.0
parameterized==0.9.0
partition==0.1.2
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.13.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
protobuf==4.25.3
py-cpuinfo==9.0.0
pycares==4.4.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pypandoc==1.13
pyparsing==3.1.1
pypdf==4.0.2
pypdfium2==4.27.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-iso639==2024.2.7
python-liquid==1.12.1
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytorchvideo==0.1.5
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
rank-bm25==0.2.2
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.3.0
tiktoken==0.7.0
timm==0.6.7
tokenizers==0.15.2
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.2
transformers==4.37.2
types-regex==2024.5.15.20240519
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unicorn==2.0.1.post1
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.13
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yacs==0.1.8
yarl==1.9.4
